# swagger-express/api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: NodeJS Mentoring Programm
  version: 1.0.0
  title: API documentation for NodeJS Mentoring Programm
  contact:
    email: ravikiran.anumula@gmail.com
host: virtserver.swaggerhub.com
basePath: /epam-node/nodejs-mentoring/1.0.0
...
schemes:
- https
- http
paths:
...
  /products:
    get:
      tags:
      - products
      summary: Get all products
      description: Get all products
      operationId: getAllProducts
      produces:
      - application/json
      parameters: []
      responses:
        200:
          description: matching criteria for search results
          schema:
            type: array
            items:
              $ref: '#/definitions/Product_response'
        500:
          description: Error
          schema:
            type: object
    post:
      tags:
      - products
      summary: Create product
      description: Create product
      operationId: createProduct
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: product details - name, price, reviews
        required: false
        schema:
          $ref: '#/definitions/Product_body'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Product_response'
        500:
          description: Error
          schema:
            type: object
    x-swagger-router-controller: products
  /products/{id}:
    get:
      tags:
      - products
      summary: Get specific product
      description: Get specific product
      operationId: getProduct
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: product ID
        required: true
        type: string
      responses:
        200:
          description: matching criteria for search results
          schema:
            $ref: '#/definitions/Product_response'
        500:
          description: Error
          schema:
            type: object
    delete:
      tags:
      - products
      summary: Delete specific product
      description: Delete specific product
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        description: product ID
        required: true
        type: string
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Error
          schema:
            type: string
    x-swagger-router-controller: products
  /producrs/{id}/reviews:
    get:
      tags:
      - products
      summary: Get specific product's reviews
      description: Get specific product's reviews
      operationId: getProductReviews
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: product ID
        required: true
        type: string
      responses:
        200:
          description: matching criteria for search results
          schema:
            type: array
            items:
              type: string
        500:
          description: Error
    x-swagger-router-controller: products
...
definitions:
...
  Product_body:
    type: object
    required:
    - name
    properties:
      name:
        type: string
        minLength: 3
      price:
        type: number
      reviews:
        type: array
        items:
          type: string
    example:
      price: 9.88
      name: name
      reviews:
      - good
      - excellent
  Product_response:
    type: object
    properties:
      name:
        type: string
      price:
        type: number
      reviews:
        type: array
        items:
          type: string
      id:
        type: string
      lastModifiedDate:
        type: string
    example:
      price: 888.9
      reviews:
      - Expensive camera mobile
      id: 5d5116c05cb5184f68fcca57
      name: Pixel 3XL
      lastModifiedDate: 2019-08-12T07:35:28.227Z
...